app/src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["nghiahm <honghia06@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
python-dotenv = "^1.0.1"
psycopg = {extras = ["binary"], version = "^3.2.3"}
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
pytest-django = "^4.9.0"
factory-boy = "^3.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 120
extend-exclude = [
    "*/migrations/*.py",
    "staticfiles/*",
]


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]
